+++ s4e-backend/src/main/java/pl/cyfronet/s4e/properties/JwtProperties.java
@@ -3,9 +3,11 @@ package pl.cyfronet.s4e.properties;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.constraints.NotEmpty;
+import java.time.Duration;
 
 @ConfigurationProperties("jwt")
 @Validated
@@ -23,4 +25,14 @@ public class JwtProperties {
 
     @NotEmpty
     private String keyPassword;
+
+    @NestedConfigurationProperty
+    private Token token = new Token();
+
+    @Getter
+    @Setter
+    public static class Token {
+        /// Half a day. In case of an 8h shift require a login every day.
+        private Duration expirationTime = Duration.ofHours(12);
+    }
 }