+++ s4e-backend/src/main/java/pl/cyfronet/s4e/security/SecurityConstants.java
@@ -1,0 +1,8 @@
+package pl.cyfronet.s4e.security;
+
+public class SecurityConstants {
+    public static final String HEADER_NAME = "Authorization";
+    public static final String COOKIE_NAME = "token";
+
+    public static final String JWT_AUTHORITIES_CLAIM = "authorities";
+}