+++ s4e-backend/src/test/java/pl/cyfronet/s4e/controller/GroupControllerTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.http.MediaType;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import pl.cyfronet.s4e.BasicTest;
@@ -29,7 +28,8 @@ import java.util.Set;
 
 import static com.github.npathai.hamcrestopt.OptionalMatchers.isPresent;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.hasSize;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -65,9 +65,6 @@ public class GroupControllerTest {
     private SlugService slugService;
 
     @Autowired
-    private PasswordEncoder passwordEncoder;
-
-    @Autowired
     private MockMvc mockMvc;
 
     private AppUser appUser;
@@ -83,7 +80,7 @@ public class GroupControllerTest {
                 .email(PROFILE_EMAIL)
                 .name("Get")
                 .surname("Profile")
-                .password(passwordEncoder.encode("password"))
+                .password("{noop}password")
                 .enabled(true)
                 .build());