+++ s4e-backend/src/test/java/pl/cyfronet/s4e/controller/InstitutionControllerTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.http.MediaType;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.web.servlet.MockMvc;
 import pl.cyfronet.s4e.BasicTest;
 import pl.cyfronet.s4e.TestDbHelper;
@@ -50,9 +49,6 @@ public class InstitutionControllerTest {
     private TestDbHelper testDbHelper;
 
     @Autowired
-    private PasswordEncoder passwordEncoder;
-
-    @Autowired
     private SlugService slugService;
 
     @Autowired
@@ -70,7 +66,7 @@ public class InstitutionControllerTest {
                 .email(PROFILE_EMAIL)
                 .name("Get")
                 .surname("Profile")
-                .password(passwordEncoder.encode("password"))
+                .password("{noop}password")
                 .enabled(true)
                 .build());
 
@@ -106,7 +102,7 @@ public class InstitutionControllerTest {
                 .email("get2@profile.com")
                 .name("Get")
                 .surname("Profile")
-                .password(passwordEncoder.encode("password"))
+                .password("{noop}password")
                 .enabled(true)
                 .build());